@@ -2,7 +2,7 @@ name: list_easy
 description: The smartest grocery list app in the world.
 publish_to: 'none'
 
-version: 1.0.1+2
+version: 1.0.2+3
 
 environment:
   sdk: '>=3.3.0 <4.0.0'
@@ -13,31 +13,36 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  # CORE
-  supabase_flutter: ^2.8.0
+  # Core & Backend
+  supabase_flutter: ^2.10.0          # Latest stable (2026)
   flutter_secure_storage: ^9.2.4
   shared_preferences: ^2.5.3
 
-  # NAVIGATION & SHARING
+  # Navigation & Sharing
   go_router: ^14.3.0
   share_plus: ^10.0.3
-  url_launcher: ^6.3.0           # Still needed for WhatsApp/Facebook sharing
+  url_launcher: ^6.3.0
 
-  # MONETIZATION
+  # Monetization
   google_mobile_ads: ^6.0.0
 
-  # PUSH NOTIFICATIONS
+  # Push Notifications
   onesignal_flutter: ^5.1.2
 
-  # UTILS
-  intl: ^0.20.2
+  # Interactive Tutorial
+  showcaseview: ^3.0.0
 
+  # Offline Detection (added for main.dart offline banner)
+  connectivity_plus: ^6.0.5
+
+  # Utils
+  intl: ^0.20.2
   cupertino_icons: ^1.0.8
 
-  # REMOVED: Google & Facebook login (we'll add them back later when it's stable)
+  # Uncomment these when adding social logins later:
   # google_sign_in: ^6.3.0
   # flutter_facebook_auth: ^7.1.1
-  # sign_in_with_apple: ^7.0.1   ← keep if you want Apple later
+  # sign_in_with_apple: ^7.0.1
 
 dev_dependencies:
   flutter_test:
@@ -57,6 +62,5 @@ flutter:
 
   assets:
     - assets/images/
-    # Removed unused social login logos
     - assets/whatsapp_logo.jpg
     - assets/WhatsApp.svg
