@@ -1,15 +1,16 @@
 // lib/pages/login/login_widget.dart
 // FINAL LAUNCH VERSION — EMAIL + PASSWORD ONLY (ROCK SOLID)
+// FIXED: Overflow on "Don't have an account?" row (yellow tape warning)
 // Friendly error messages — no more scary AuthExceptions
-// Graceful offline handling with nice message
+// UPDATED: Clean UI, no offline mode, prominent Sign Up link, keyboard fix
 
-import 'dart:io'; // For SocketException
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 class LoginWidget extends StatefulWidget {
   const LoginWidget({super.key});
+
   @override
   State<LoginWidget> createState() => _LoginWidgetState();
 }
@@ -32,28 +33,19 @@ class _LoginWidgetState extends State<LoginWidget> {
       // Success → go to create list
       if (mounted) context.go('/create');
     } on AuthException catch (e) {
-      String message = 'Please sign up first';
+      String message = 'Login failed — please try again';
 
       if (e.message.toLowerCase().contains('invalid login credentials') ||
           e.message.toLowerCase().contains('user not found')) {
-        message = 'Please sign up first';
+        message = 'Email not found — try signing up';
       } else if (e.message.toLowerCase().contains('password')) {
         message = 'Incorrect password';
       }
 
-      if (mounted) {
-        _showError(message);
-      }
-    } on SocketException catch (_) {
-      // No internet
-      if (mounted) {
-        _showError(
-            'No internet connection — please check your network and try again');
-      }
+      if (mounted) _showError(message);
     } catch (e) {
-      // Any other error
       if (mounted) {
-        _showError('Login failed — please try again');
+        _showError('Something went wrong — please try again');
       }
     } finally {
       if (mounted) setState(() => isLoading = false);
@@ -65,7 +57,7 @@ class _LoginWidgetState extends State<LoginWidget> {
       SnackBar(
         content: Text(message),
         backgroundColor: Colors.deepOrange,
-        duration: const Duration(seconds: 6),
+        duration: const Duration(seconds: 5),
       ),
     );
   }
@@ -74,74 +66,104 @@ class _LoginWidgetState extends State<LoginWidget> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.black,
+      resizeToAvoidBottomInset: true, // Keyboard scrolls content up
       body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.all(32),
+        child: SingleChildScrollView(
+          padding: EdgeInsets.only(
+            left: 32,
+            right: 32,
+            top: 40,
+            bottom: MediaQuery.of(context).viewInsets.bottom + 40,
+          ),
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.center,
             children: [
+              // Welcome header
               const Text(
-                'Welcome to List Easy',
+                'Welcome Back!',
                 style: TextStyle(
-                  fontSize: 28,
+                  fontSize: 32,
                   fontWeight: FontWeight.bold,
                   color: Colors.white,
                 ),
+                textAlign: TextAlign.center,
               ),
-              const SizedBox(height: 60),
+              const SizedBox(height: 12),
+              const Text(
+                'Log in to continue organizing your shopping lists.',
+                style: TextStyle(
+                  fontSize: 16,
+                  color: Colors.white70,
+                  height: 1.4,
+                ),
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 48),
 
-              // Email
+              // Form fields
               TextField(
                 controller: emailController,
                 style: const TextStyle(color: Colors.white),
+                decoration: _inputDecoration('Email Address'),
                 keyboardType: TextInputType.emailAddress,
-                decoration: _input('Email'),
+                autocorrect: false,
               ),
-              const SizedBox(height: 16),
-
-              // Password
+              const SizedBox(height: 20),
               TextField(
                 controller: passwordController,
                 obscureText: true,
                 style: const TextStyle(color: Colors.white),
-                decoration: _input('Password'),
+                decoration: _inputDecoration('Password'),
               ),
-              const SizedBox(height: 32),
+              const SizedBox(height: 40),
 
-              // Sign In Button
+              // Sign In button
               SizedBox(
                 width: double.infinity,
                 height: 56,
                 child: ElevatedButton(
                   onPressed: isLoading ? null : _loginWithEmail,
                   style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.white,
+                    backgroundColor: Colors.cyan,
+                    foregroundColor: Colors.black,
                     shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(30),
                     ),
+                    elevation: 2,
                   ),
                   child: isLoading
-                      ? const CircularProgressIndicator(color: Colors.black)
+                      ? const SizedBox(
+                          height: 24,
+                          width: 24,
+                          child: CircularProgressIndicator(
+                            color: Colors.black,
+                            strokeWidth: 3,
+                          ),
+                        )
                       : const Text(
                           'Sign In',
                           style: TextStyle(
-                            color: Colors.black,
-                            fontWeight: FontWeight.bold,
                             fontSize: 18,
+                            fontWeight: FontWeight.bold,
                           ),
                         ),
                 ),
               ),
 
-              const SizedBox(height: 60),
+              const SizedBox(height: 32),
 
-              // Sign Up link
+              // Don't have account? Sign Up — FIXED overflow
               Row(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                  const Text(
-                    "Don't have an account? ",
-                    style: TextStyle(color: Colors.white70),
+                  Flexible(
+                    child: Text(
+                      "Don't have an account? ",
+                      style:
+                          const TextStyle(color: Colors.white70, fontSize: 16),
+                      overflow: TextOverflow.ellipsis,
+                      maxLines: 1,
+                    ),
                   ),
                   GestureDetector(
                     onTap: () => context.go('/signup'),
@@ -150,10 +172,23 @@ class _LoginWidgetState extends State<LoginWidget> {
                       style: TextStyle(
                         color: Colors.cyan,
                         fontWeight: FontWeight.bold,
+                        fontSize: 16,
                       ),
                     ),
                   ),
                 ],
+              ),
+
+              const SizedBox(height: 48),
+
+              // Built with Grok credit
+              const Text(
+                'Built with Grok by xAI',
+                style: TextStyle(
+                  fontSize: 12,
+                  color: Colors.white54,
+                  fontWeight: FontWeight.w300,
+                ),
               ),
             ],
           ),
@@ -162,16 +197,24 @@ class _LoginWidgetState extends State<LoginWidget> {
     );
   }
 
-  InputDecoration _input(String hint) => InputDecoration(
-        hintText: hint,
-        hintStyle: const TextStyle(color: Colors.white54),
-        filled: true,
-        fillColor: const Color(0xFF1A1A1A),
-        border: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(30),
-          borderSide: BorderSide.none,
-        ),
-        contentPadding:
-            const EdgeInsets.symmetric(horizontal: 20, vertical: 18),
-      );
+  InputDecoration _inputDecoration(String hint) {
+    return InputDecoration(
+      hintText: hint,
+      hintStyle: const TextStyle(color: Colors.white54),
+      filled: true,
+      fillColor: const Color(0xFF1E1E1E),
+      border: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(30),
+        borderSide: BorderSide.none,
+      ),
+      contentPadding: const EdgeInsets.symmetric(horizontal: 24, vertical: 18),
+    );
+  }
+
+  @override
+  void dispose() {
+    emailController.dispose();
+    passwordController.dispose();
+    super.dispose();
+  }
 }
