@@ -1,14 +1,16 @@
 // lib/pages/signup/signup_widget.dart
 // FINAL LAUNCH VERSION — CLEAN SIGN-UP WITH FRIENDLY ERRORS
+// FIXED: Overflow on "Already have an account?" row (yellow tape warning)
 // Graceful offline handling + KEYBOARD FIX (content scrolls above keyboard)
+// UPDATED: First-time user focus, no offline mode, welcoming UI, prominent Sign In link
 
-import 'dart:io'; // For SocketException
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 class SignupWidget extends StatefulWidget {
   const SignupWidget({super.key});
+
   @override
   State<SignupWidget> createState() => _SignupWidgetState();
 }
@@ -57,16 +59,9 @@ class _SignupWidgetState extends State<SignupWidget> {
       }
 
       if (mounted) _showError(message);
-    } on SocketException catch (_) {
-      // No internet
-      if (mounted) {
-        _showError(
-            'No internet connection — please check your network and try again');
-      }
     } catch (e) {
-      // Any other unexpected error
       if (mounted) {
-        _showError('Sign-up failed — please try again');
+        _showError('Something went wrong — please try again');
       }
     } finally {
       if (mounted) setState(() => isLoading = false);
@@ -78,7 +73,7 @@ class _SignupWidgetState extends State<SignupWidget> {
       SnackBar(
         content: Text(message),
         backgroundColor: Colors.deepOrange,
-        duration: const Duration(seconds: 6),
+        duration: const Duration(seconds: 5),
       ),
     );
   }
@@ -87,103 +82,173 @@ class _SignupWidgetState extends State<SignupWidget> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.black,
-      resizeToAvoidBottomInset: true, // Important for keyboard
+      resizeToAvoidBottomInset: true, // Keyboard scrolls content up
       body: SafeArea(
-        child: Padding(
+        child: SingleChildScrollView(
           padding: EdgeInsets.only(
             left: 32,
             right: 32,
-            top: 32,
-            bottom: MediaQuery.of(context).viewInsets.bottom +
-                32, // Magic: pushes content above keyboard
+            top: 40,
+            bottom: MediaQuery.of(context).viewInsets.bottom + 40,
           ),
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const Text('Create Account',
-                    style: TextStyle(
-                        fontSize: 32,
-                        fontWeight: FontWeight.bold,
-                        color: Colors.white)),
-                const SizedBox(height: 40),
-                TextField(
-                  controller: nameController,
-                  style: const TextStyle(color: Colors.white),
-                  decoration: _input('Full Name'),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              // Welcome header
+              const Text(
+                'Welcome to List Easy! 🎉',
+                style: TextStyle(
+                  fontSize: 32,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.white,
                 ),
-                const SizedBox(height: 16),
-                TextField(
-                  controller: emailController,
-                  style: const TextStyle(color: Colors.white),
-                  decoration: _input('Email Address'),
-                  keyboardType: TextInputType.emailAddress,
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 12),
+              const Text(
+                'Create your account and start organizing your shopping lists like never before.',
+                style: TextStyle(
+                  fontSize: 16,
+                  color: Colors.white70,
+                  height: 1.4,
                 ),
-                const SizedBox(height: 16),
-                TextField(
-                  controller: passwordController,
-                  obscureText: true,
-                  style: const TextStyle(color: Colors.white),
-                  decoration: _input('Password'),
-                ),
-                const SizedBox(height: 16),
-                TextField(
-                  controller: confirmController,
-                  obscureText: true,
-                  style: const TextStyle(color: Colors.white),
-                  decoration: _input('Confirm Password'),
-                ),
-                const SizedBox(height: 32),
-                SizedBox(
-                  width: double.infinity,
-                  height: 56,
-                  child: ElevatedButton(
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.white,
-                      shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(30)),
+                textAlign: TextAlign.center,
+              ),
+              const SizedBox(height: 48),
+
+              // Form fields
+              TextField(
+                controller: nameController,
+                style: const TextStyle(color: Colors.white),
+                decoration: _inputDecoration('Full Name'),
+                textCapitalization: TextCapitalization.words,
+              ),
+              const SizedBox(height: 20),
+              TextField(
+                controller: emailController,
+                style: const TextStyle(color: Colors.white),
+                decoration: _inputDecoration('Email Address'),
+                keyboardType: TextInputType.emailAddress,
+                autocorrect: false,
+              ),
+              const SizedBox(height: 20),
+              TextField(
+                controller: passwordController,
+                obscureText: true,
+                style: const TextStyle(color: Colors.white),
+                decoration: _inputDecoration('Password'),
+              ),
+              const SizedBox(height: 20),
+              TextField(
+                controller: confirmController,
+                obscureText: true,
+                style: const TextStyle(color: Colors.white),
+                decoration: _inputDecoration('Confirm Password'),
+              ),
+              const SizedBox(height: 40),
+
+              // Sign Up button
+              SizedBox(
+                width: double.infinity,
+                height: 56,
+                child: ElevatedButton(
+                  onPressed: isLoading ? null : _signup,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: Colors.cyan,
+                    foregroundColor: Colors.black,
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(30),
                     ),
-                    onPressed: isLoading ? null : _signup,
-                    child: isLoading
-                        ? const CircularProgressIndicator(color: Colors.black)
-                        : const Text('Sign Up',
-                            style: TextStyle(
-                                color: Colors.black,
-                                fontWeight: FontWeight.bold,
-                                fontSize: 18)),
+                    elevation: 2,
                   ),
-                ),
-                const SizedBox(height: 40),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    const Text('Already have an account? ',
-                        style: TextStyle(color: Colors.white70)),
-                    GestureDetector(
-                      onTap: () => context.go('/login'),
-                      child: const Text('Sign In',
+                  child: isLoading
+                      ? const SizedBox(
+                          height: 24,
+                          width: 24,
+                          child: CircularProgressIndicator(
+                            color: Colors.black,
+                            strokeWidth: 3,
+                          ),
+                        )
+                      : const Text(
+                          'Create Account',
                           style: TextStyle(
-                              color: Colors.cyan, fontWeight: FontWeight.bold)),
-                    ),
-                  ],
+                            fontSize: 18,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
                 ),
-              ],
-            ),
+              ),
+
+              const SizedBox(height: 32),
+
+              // Already have account? Sign In — FIXED overflow
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Flexible(
+                    child: Text(
+                      'Already have an account? ',
+                      style: const TextStyle(
+                        color: Colors.white70,
+                        fontSize: 16,
+                      ),
+                      overflow: TextOverflow.ellipsis,
+                      maxLines: 1,
+                    ),
+                  ),
+                  GestureDetector(
+                    onTap: () => context.go('/login'),
+                    child: const Text(
+                      'Sign In',
+                      style: TextStyle(
+                        color: Colors.cyan,
+                        fontWeight: FontWeight.bold,
+                        fontSize: 16,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+
+              const SizedBox(height: 48),
+
+              // Built with Grok credit
+              const Text(
+                'Built with Grok by xAI',
+                style: TextStyle(
+                  fontSize: 12,
+                  color: Colors.white54,
+                  fontWeight: FontWeight.w300,
+                ),
+              ),
+            ],
           ),
         ),
       ),
     );
   }
 
-  InputDecoration _input(String hint) => InputDecoration(
-        hintText: hint,
-        hintStyle: const TextStyle(color: Colors.white54),
-        filled: true,
-        fillColor: const Color(0xFF1A1A1A),
-        border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(30),
-            borderSide: BorderSide.none),
-        contentPadding:
-            const EdgeInsets.symmetric(horizontal: 20, vertical: 18),
-      );
+  InputDecoration _inputDecoration(String hint) {
+    return InputDecoration(
+      hintText: hint,
+      hintStyle: const TextStyle(color: Colors.white54),
+      filled: true,
+      fillColor: const Color(0xFF1E1E1E),
+      border: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(30),
+        borderSide: BorderSide.none,
+      ),
+      contentPadding: const EdgeInsets.symmetric(horizontal: 24, vertical: 18),
+    );
+  }
+
+  @override
+  void dispose() {
+    nameController.dispose();
+    emailController.dispose();
+    passwordController.dispose();
+    confirmController.dispose();
+    super.dispose();
+  }
 }
